move get_event_stream to utils
