+++ producer/utils.py
@@ -104,3 +104,12 @@ def get_all_csc_names_in_message(message):
 
 
 
+def get_event_stream(message, category, csc, salindex, stream_name):
+    """ Tries to return the first stream found in a LOVE message. 
+    Throws errors if it does not exist. """
+
+    data_generator = (d for d in message['data'] if d["csc"] == csc and d['salindex'] == salindex)
+    data = next(data_generator)
+    stream_generator = (data['data'][s] for s in data['data'] if s == stream_name)
+    stream = next(stream_generator)
+    return stream
